requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Database
psycopg2-binary>=2.9.0

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipython>=8.0.0
notebook>=7.0.0

# Utilities
python-dateutil>=2.8.0
